requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
httpx==1.0.0
